requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.1.7
django-autoslug==1.9.9
django-resized==1.0.2
gunicorn==20.1.0
idna==3.4
joblib==1.2.0
nltk==3.8.1
Pillow==9.4.0
pytz==2023.3
regex==2023.3.23
requests==2.28.2
soupsieve==2.4.1
sqlparse==0.4.3
tqdm==4.65.0
tzdata==2022.7
urllib3==1.26.15
whitenoise==6.4.0
